+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 - removed stray pixels in modular clouds icon
 - added an option to display relative humidity in the overview chart
 - added warnings filter to overview chart
-- added background location check call to onResume
 - fixed main activity terminating twice under certain conditions
+- added hint in settings when background location checks are enabled but the app is lacking the background location permission
+- added background location check call to onResume
+- a new background location is now also loaded in onResume
 
 ## Version 0.62.3:
 - add monochrome app icon to support app icon theming